requirements.txt
matplotlib==3.10.0
mysql-connector-python
numpy==2.2.2
pandas==2.2.3
pymongo==4.10.1
streamlit==1.41.1
